Convert the last attribute into tags and count the number of dropped attributes, when attributes with a duplicate key are set (#17)
OTEL's spec technically allows attributes to contain multiple same keys
(I actually encountered this use case). If converted as is, an
`IllegalStateException` occurs.
In this PR, if multiple same keys are added, the attribute will be used.
This is the same behavior as when creating a Brave span. And dropped
attributes are counted.