Remove dependency on io.opentelemetry.semconv:opentelemetry-semconv (#23)
This pull request removes the dependency on opentelemetry-semconv.
This jar is alpha prefixed and may be subject to future breaking
changes. To prevent changes to semconv from breaking brave on the client
side, I have made semconv attributes constant and have them on the brave
side.
In fact, semconv is only used for converting to and from brace tags, so
the benefit of having this dependency is very small compared to the
risk.

With this change, the only dependency on otel is proto.